+++ test/unit/mysql_rep_failover/create_instances.py
@@ -30,7 +30,6 @@ import mock
 # Local
 sys.path.append(os.getcwd())
 import mysql_rep_failover
-import lib.gen_libs as gen_libs
 import version
 
 __version__ = version.__version__
@@ -64,6 +63,8 @@ class SlaveRep(object):
         self.name = name
         self.exe_gtidset = exe_gtidset
         self.gtid_mode = gtid_mode
+        self.master = None
+        self.slave = None
 
     def remove(self, master):
 
@@ -75,6 +76,8 @@ class SlaveRep(object):
             (input) master -> Master name.
 
         """
+
+        self.master = master
 
         return True
 
@@ -88,6 +91,8 @@ class SlaveRep(object):
             (input) slave -> Slave name.
 
         """
+
+        self.slave = slave
 
         return True